refactor: separa config do formulário login em um hook
